refactor benefit intake client logic into its own class
